\name{gig_cpp}
\alias{gig_cpp}
\title{Rgig wrapper for the c++ code simulating gig}
\usage{
gig_cpp(p, a, b)
}
\arguments{
  \item{a}{- (nx1 - double ) input parameter for gig}

  \item{b}{- (nx1 - double ) input parameter for gig}

  \item{p}{- (nx1 - double) input parameter for gig}
}
\value{
(nx1 - double)
}
\description{
Rgig wrapper for the c++ code simulating gig
}

